+++ libffi-rs/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog] and this project adheres to
   * Deprecated many functions in `middle::Type`, preferring the usage of enum variants where
     possible, but it is still possible to use the old functions for backwards compatibility reasons.
 - Removed libc dependency
+- Made crates no_std
 
 ## [3.2.0] - 2023-03-28